+++ NEWS.md
@@ -2,9 +2,10 @@
 
 ### Enhancements
  * Analysis functions (`cfun/afun`) have more information about current column split; `.spl_context` has access to `cur_col_id`, `cur_col_expr`, `cur_col_split`, and `cur_col_split_val`.
- * Analysis functions (`cfun/afun`) can use new parameters to extend analysis calculations. `.alt_df` gives access to `alt_counts_df` across columns, while `.all_col_exprs` and `.all_col_counts` contains global information about all columns.
- * Added check for `.alt_df` that prevents its calculation if not present in analysis functions.
- * 
+ * Analysis functions (`cfun/afun`) can use new parameters to extend analysis calculations. `.alt_df_row` gives access to `alt_counts_df` across columns, while `.all_col_exprs` and `.all_col_counts` contains global information about all columns.
+ * Added `.alt_df` to have the exact same subset of `df` in `afun/cfun`.
+ * Added check for `.alt_df_row` that prevents its calculation if not present in analysis functions.
+
 ## rtables 0.6.2
  * Fixed major regressions for `page_by` machinery caused by migration to `formatters` 0.5.1 pagination framework.
  * Fixed `page_by` labels become missing when only one level exist in the `split_rows_by`.